@@ -34,8 +34,9 @@
     "prettier:check": "prettier --check \"{src,example}/**/*.{js,ts,tsx}\"",
     "prettier:write": "prettier --write \"{src,example}/**/*.{js,ts,tsx}\"",
     "pretest": "cd example && yarn",
-    "test": "yarn storybook:build && concurrently -k -s first -n \"SB, TEST\" -c \"magenta,blue\" \"yarn storybook:preview\" \"yarn storybook:test-ci\"",
-    "test:unit": "cross-env CI=1 jest src --env=jsdom",
+    "test": "yarn test:unit && yarn test:sb",
+    "test:sb": "yarn storybook:build && concurrently -k -s first -n \"SB, TEST\" -c \"magenta,blue\" \"yarn storybook:preview\" \"yarn storybook:test-ci\"",
+    "test:unit": "cross-env CI=1 jest src --env=jsdom --colors --silent",
     "test:watch": "jest src --watchAll --env=jsdom",
     "hooks:uninstall": "husky uninstall",
     "hooks:install": "husky install",
@@ -156,11 +157,13 @@
     "http-server": "14.1.1",
     "husky": "8.0.3",
     "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
+    "jsdom": "22.1.0",
     "npm-run-all": "4.1.5",
     "pinst": "3.0.0",
     "prettier": "3.0.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-i18next": "13.2.0",
     "react-query": "3.39.3",
     "react-router-dom": "6.16.0",
