@@ -5,18 +5,14 @@ module.exports = {
     '^.+\\.(js|jsx)$': 'babel-jest',
   },
   moduleNameMapper: {
-    '@material-ui/core/(.*)':
-      '<rootDir>/example/node_modules/@material-ui/core/$1',
-    '@material-ui/icons/(.*)':
-      '<rootDir>/example/node_modules/@material-ui/icons/$1',
-    '@material-ui/lab/(.*)':
-      '<rootDir>/example/node_modules/@material-ui/lab/$1',
-    '@material-ui/styles': '<rootDir>/example/node_modules/@material-ui/styles',
+    '@mui/material/(.*)': '<rootDir>/node_modules/@mui/material/$1',
+    '@mui/icons-material/(.*)': '<rootDir>/node_modules/@mui/icons-material/$1',
+    '@mui/lab/(.*)': '<rootDir>/node_modules/@mui/lab/$1',
     immutable: '<rootDir>/node_modules/immutable',
-    '^react$': '<rootDir>/example/node_modules/react',
-    '^react-router-dom$': '<rootDir>/example/node_modules/react-router-dom',
-    '^react-dom$': '<rootDir>/example/node_modules/react-dom',
-    '^react-dom/(.*)': '<rootDir>/example/node_modules/react-dom/$1',
+    '^react$': '<rootDir>/node_modules/react',
+    '^react-router-dom$': '<rootDir>/node_modules/react-router-dom',
+    '^react-dom$': '<rootDir>/node_modules/react-dom',
+    '^react-dom/(.*)': '<rootDir>/node_modules/react-dom/$1',
     '\\.(css|less)$': '<rootDir>/test/styleMock.js',
   },
   resolver: `${__dirname}/test/resolver.js`,
